requirements.txt
streamlit
pymysql
pandas
python-dotenv
playwright
openpyxl
